requirements.txt
python-telegram-bot
pysqlite3